travis: Switch to dpdk main branch.
Make this branch point to current main master branch so that we can
track API breakage.

Note: this should not be merged to master, intended for dpdk-latest
branch only.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>

The default branch name in DPDK is changed from master to main.
This patch reflects the same on travis builds for dpdk-latest branch.

Tested-at: https://travis-ci.org/github/Sunil-Pai-G/ovs-copy/builds/723223426
Co-authored-by: Sunil Pai G <sunil.pai.g@intel.com>
Signed-off-by: Sunil Pai G <sunil.pai.g@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>