ci: Switch to dpdk main branch.
Make this branch point to current DPDK main branch so that we can
track API breakage for GitHub Actions.

In GitHub Actions, debian packaging jobs rely on a packaged version of
dpdk and thus are not relevant.

Note: this should not be merged to master, intended for dpdk-latest
branch only.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Sunil Pai G <sunil.pai.g@intel.com>
Co-authored-by: Sunil Pai G <sunil.pai.g@intel.com>
Signed-off-by: istokes <ian.stokes@intel.com>