dpif-netdev: Fix segfault after port delete
Before this fix a segfault can occur after deleting a port. If packets
received on the deleted port are still in flight on a different async
vhost port, then the RXQ of the deleted port can be cleaned up before
all packets received on that port have been fully processed.

Ideally the cleanup of the RXQs associated with a deleted port would be
delayed till after all packets received on the port have been fully
processed. Checking whether the RXQ pointer is valid will also fix this
issue. After the async TX has been started, the RXQ pointer is only used
for cycle counting, so the check below is enough.

Signed-off-by: Cian Ferriter <cian.ferriter@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>